Added normalized name to publication_contributors, split views_and_indices into two separate files
