+++ dataprocessing/views.py
@@ -1,5 +1,4 @@
 import sqlite3
-import csv_db
 
 def train_test_views(dataset=''):
 	"""Create database views for train and test set of authorship_ggc table
@@ -16,25 +15,7 @@ def train_test_views(dataset=''):
 		statements[view_name] = statement
 	return statements
 
-def create_index(table, columns,db = '../data/demosaurus.sqlite'):
-	"""General function for creating indices on specified table for specified columns
-	"""
-	statement = "CREATE INDEX IF NOT EXISTS %s" % '_'.join([table]+columns)
-	statement += "\nON %s(%s)" % (table, ','.join(columns))
-
-	with sqlite3.connect(db) as con:
-		c = con.cursor()
-		c.execute(statement)	
-
-def set_fts5(db='../data/demosaurus.sqlite'):
-	with sqlite3.connect(db) as con:
-		# Create virtual author table for high-performance text search (obtain candidates)
-		con.execute("DROP TABLE IF EXISTS `author_fts5`;")
-		con.execute("CREATE VIRTUAL TABLE author_fts5 USING FTS5(author_ppn, searchkey, name, name_normalized, familyname);")
-		con.execute("INSERT INTO author_fts5 SELECT author_ppn, searchkey, name, name_normalized, familyname FROM author_name_options ;")
-	return
-
-def author_aggregated_query():
+def author_aggregated_query(ignore = []):
 	feature_specs = {'CBK_genre':{},
 					 'NUGI_genre':{},
 					 'NUR_rubriek':{},
@@ -45,10 +26,12 @@ def author_aggregated_query():
 					 'jaar_van_uitgave': {'table_name':'publication_basicinfo', 'feature_column': 'jaar_van_uitgave'},
 					 'role':{'table_name':'publication_contributors', 'feature_column':'role'}
 					 }
+	for feaure in ignore:
+		del feature_specs[feaure]
 	query = f"""WITH publication_subset AS(
 			SELECT t0.publication_ppn, t0.author_ppn
 			FROM publication_contributors_train_NBD t0
-			WHERE t0.author_ppn	IS NOT NULL)"""
+			WHERE t0.author_ppn IS NOT NULL)"""
 	for i, (feature_name, specs) in enumerate(feature_specs.items()):
 		if i>0: query += "\nUNION"
 		table_name = specs['table_name'] if 'table_name' in specs else 'publication_'+feature_name
@@ -82,13 +65,3 @@ def create_views(db = '../data/demosaurus.sqlite'):
 			print(view_name, statement)
 			c.execute(statement)
 
-def create_indices(db = '../data/demosaurus.sqlite'):
-
-	for table_name in csv_db.schemata.keys():
-		if 'publication_' in table_name:
-			create_index(table_name, ['publication_ppn'], db=db)  # Common entry: look up by publication
-		if 'author_' in table_name:
-			create_index(table_name, ['author_ppn'], db=db)  # Common entry: look up by author
-	create_index('publication_contributors', ['author_ppn'],
-				 db=db)  # Common entry for authorship: look up publications by author
-	set_fts5(db=db) # To speed-up retrieval of candidates
